Updated Controllers mappings namings, Cruds and thier return types + services and mappers

+++ src/main/kotlin/codeacademy/bookingforum/app/photo/PhotoService.java
@@ -1,5 +1,7 @@
 package codeacademy.bookingforum.app.photo;
 
+import codeacademy.bookingforum.app.post.Post;
+import codeacademy.bookingforum.app.user.auth.UserAuth;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,34 +16,48 @@ public class PhotoService {
     @Autowired
     PhotoMapper photoMapper;
 
-    public PhotoDto findById(Long id){
+    public PhotoDto findById(Long id) {
         Photo photo = photoRepo.findById(id).orElse(null);
         return photoMapper.toDto(photo);
     }
 
-    public List<PhotoDto> findAllPhotos(){
+    public List<PhotoDto> findAllPhotos() {
         List<Photo> photos = (List<Photo>) photoRepo.findAll();
-        return  photoMapper.toDto(photos);
+        return photoMapper.toDto(photos);
     }
 
 
-    public PhotoDto createPhoto(PhotoDto photoDto){
+    public PhotoDto createPhoto(PhotoDto photoDto) {
         Photo photo = photoMapper.fromDto(photoDto);
         photoRepo.save(photo);
         return photoMapper.toDto(photo);
     }
 
-    public PhotoDto updatePhoto(PhotoDto photoDto){
-        Photo photo = photoMapper.fromDto(photoDto);
-        photoRepo.save(photo);
-        return photoMapper.toDto(photo);
+    public String updatePhoto(Long id, PhotoDto photoDto) {
+        if(photoRepo.existsById(id)) {
+            photoRepo.save(
+                    new Photo(
+                            id,
+                            photoDto.getImageUrl(),
+                            photoDto.getTags(),
+                            photoDto.getDescription(),
+                            photoDto.getType(),
+                            photoDto.getUserPhotos_id().stream().map( ids -> new UserAuth(ids) ).toList(),
+                            photoDto.getPostPhoto_id().stream().map( ids -> new Post(ids)).toList()));
+            return "Photo with id " + id + " updated";
+        }else {
+            return "Photo with id " + id + " not found";
+        }
     }
 
-    public void deletePhoto(Long id){
-        photoRepo.deleteById(id);
+    public String deletePhoto(Long id) {
+        if (photoRepo.existsById(id)) {
+            photoRepo.deleteById(id);
+            return "Photo with id " + id + " deleted";
+        } else {
+            return "Photo with id " + id + " does not exist";
+        }
     }
-
-
 
 
 }
